requirements.txt
Flask==1.1.2
influxdb==5.3.0
numpy==1.19.1
pandas==1.1.0
PyYAML==5.3.1
requests==2.24.0

plotly==4.9.0